+++ qtgl/projection.hpp
@@ -11,54 +11,54 @@ enum class GLProjectionMode {
 
 class GLProjection {
  public:
-  float height, width;
-  float near, far;
+  double height, width;
+  double near, far;
   GLProjectionMode mode;
   GLProjection()
       : height(768), width(1024), near(0.1f), far(100), mode(GLProjectionMode::PRESPECTIVE) {}
-  GLProjection(float height, float width, float near, float far, GLProjectionMode mode)
+  GLProjection(double height, double width, double near, double far, GLProjectionMode mode)
       : height(height), width(width), near(near), far(far), mode(mode) {}
-  Eigen::Matrix4f orthographicProjMatrix() {
-    float hfov = MathUtils::PI / 3;
-    float vfov = hfov * (height / width);
-    float right = tan(hfov / 2) * near;
-    float left = -right;
-    float top = tan(vfov / 2) * near;
-    float bottom = -top;
+  Eigen::Matrix4d orthographicProjMatrix() {
+    double hfov = MathUtils::PI / 3;
+    double vfov = hfov * (height / width);
+    double right = tan(hfov / 2) * near;
+    double left = -right;
+    double top = tan(vfov / 2) * near;
+    double bottom = -top;
 
-    float m00 = 2 / (right - left);
-    float m11 = 2 / (top - bottom);
-    float m22 = 2 / (near - far);
-    float m30 = -(right + left) / (right - left);
-    float m31 = -(top + bottom) / (top - bottom);
-    float m32 = -(near + far) / (near - far);
+    double m00 = 2 / (right - left);
+    double m11 = 2 / (top - bottom);
+    double m22 = 2 / (near - far);
+    double m30 = -(right + left) / (right - left);
+    double m31 = -(top + bottom) / (top - bottom);
+    double m32 = -(near + far) / (near - far);
 
-    Eigen::Matrix4f projMtx;
+    Eigen::Matrix4d projMtx;
     projMtx << m00, 0, 0, 0,  //
         0, m11, 0, 0,         //
         0, 0, m22, 0,         //
         m30, m31, m32, 1;
     return projMtx;
   }
-  Eigen::Matrix4f perspectiveProjMatrix() {
-    float hfov = MathUtils::PI / 3;
-    float vfov = hfov * (height / width);
-    float right = tan(hfov / 2) * near;
-    float left = -right;
-    float top = tan(vfov / 2) * near;
-    float bottom = -top;
-    float m00 = 2 / (right - left);
-    float m11 = 2 / (top - bottom);
-    float m22 = (far + near) / (far - near);
-    float m32 = -2 * near * far / (far - near);
-    Eigen::Matrix4f projMtx;
+  Eigen::Matrix4d perspectiveProjMatrix() {
+    double hfov = MathUtils::PI / 3;
+    double vfov = hfov * (height / width);
+    double right = tan(hfov / 2) * near;
+    double left = -right;
+    double top = tan(vfov / 2) * near;
+    double bottom = -top;
+    double m00 = 2 / (right - left);
+    double m11 = 2 / (top - bottom);
+    double m22 = (far + near) / (far - near);
+    double m32 = -2 * near * far / (far - near);
+    Eigen::Matrix4d projMtx;
     projMtx << m00, 0, 0, 0,  //
         0, m11, 0, 0,         //
         0, 0, m22, 1,         //
         0, 0, m32, 0;
     return projMtx;
   }
-  Eigen::Matrix4f projMatrix() {
+  Eigen::Matrix4d projMatrix() {
     if (mode == GLProjectionMode::ORTHOGRAPHIC) {
       return orthographicProjMatrix();
     } else {
