+++ qtgl/shader.hpp
@@ -6,18 +6,18 @@ namespace qtgl {
 
 struct GLLight {
   Color01 intensity;
-  virtual Eigen::Vector3f uvLight(Position3& pos) = 0;  // l unit vector
+  virtual Eigen::Vector3d uvLight(Position3& pos) = 0;  // l unit vector
 };
 
 struct DirectionalGLLight : public GLLight {
-  Eigen::Vector3f d;  // direction
-  Eigen::Vector3f uvLight(Position3& pos) { return (d * -1).normalized(); }
+  Eigen::Vector3d d;  // direction
+  Eigen::Vector3d uvLight(Position3& pos) { return (d * -1).normalized(); }
 };
 
 struct PointGLLight : public GLLight {
   Position3 position;
-  Eigen::Vector3f uvLight(Position3& pos) {
-    Eigen::Vector3f v{position.x - pos.x, position.y - pos.y, position.z - pos.z};
+  Eigen::Vector3d uvLight(Position3& pos) {
+    Eigen::Vector3d v{position.x - pos.x, position.y - pos.y, position.z - pos.z};
     return v.normalized();
   }
 };
@@ -25,16 +25,16 @@ struct PointGLLight : public GLLight {
 struct GLShader {};
 struct LambertianPhongGLShader : public GLShader {
  private:
-  float phongExp = 1;  // Phong Exponent
+  double phongExp = 1;  // Phong Exponent
 
  public:
-  void setPhoneExp(float p) { this->phongExp = p; }
-  float getPhoneExp() const { return this->phongExp; }
-  Color01 shade(GLLight* lighter, Eigen::Vector3f& uvNormal, Position3& pos, Eigen::Vector3f& uvEye,
+  void setPhoneExp(double p) { this->phongExp = p; }
+  double getPhoneExp() const { return this->phongExp; }
+  Color01 shade(GLLight* lighter, Eigen::Vector3d& uvNormal, Position3& pos, Eigen::Vector3d& uvEye,
                 Color01& diffuse, Color01& ambient) {
-    Eigen::Vector3f uvLight = lighter->uvLight(pos);
-    Eigen::Vector3f uvHalf = (uvEye + uvLight).normalized();  // unit halfway vector
-    float cp = 1 - std::max(diffuse.R, std::max(diffuse.G, diffuse.B));
+    Eigen::Vector3d uvLight = lighter->uvLight(pos);
+    Eigen::Vector3d uvHalf = (uvEye + uvLight).normalized();  // unit halfway vector
+    double cp = 1 - std::max(diffuse.R, std::max(diffuse.G, diffuse.B));
     Color01 c;
     c.R = shadeComponent(diffuse.R, ambient.R, lighter->intensity.R, uvNormal, uvLight, cp, uvHalf,
                          phongExp);
@@ -45,9 +45,10 @@ struct LambertianPhongGLShader : public GLShader {
     c.clamp();  // clamp 0 - 1
     return c;
   }
-  inline float shadeComponent(float cr, float ca, float cl, Eigen::Vector3f& n, Eigen::Vector3f& l,
-                              float cp, Eigen::Vector3f& h, float p) noexcept {
-    return cr * (ca + cl * std::max(0.0f, n.dot(l))) + cl * cp * std::pow(h.dot(n), p);
+  inline double shadeComponent(double cr, double ca, double cl, Eigen::Vector3d& n,
+                               Eigen::Vector3d& l, double cp, Eigen::Vector3d& h,
+                               double p) noexcept {
+    return cr * (ca + cl * std::max(0.0, n.dot(l))) + cl * cp * std::pow(h.dot(n), p);
   };
 };
 }  // namespace qtgl
